move naming conversation to conversation creation
