+++ app/api/conversations/route.ts
@@ -1,15 +1,20 @@
 import assert from "assert";
 
+import { openai } from "@ai-sdk/openai";
+import { generateText } from "ai";
 import { and, desc, eq, sql } from "drizzle-orm";
 import { NextRequest } from "next/server";
 import { z } from "zod";
 
+import { NAMING_SYSTEM_PROMPT } from "@/lib/constants";
+import { DEFAULT_NAMING_MODEL } from "@/lib/llm/types";
 import db from "@/lib/server/db";
 import * as schema from "@/lib/server/db/schema";
 import { requireAuthContextFromRequest } from "@/lib/server/utils";
 
+
 const createConversationRequest = z.object({
-  title: z.string(),
+  content: z.string(),
 });
 
 const getConversationsRequest = z.object({
@@ -20,7 +25,8 @@ const getConversationsRequest = z.object({
 export async function POST(request: NextRequest) {
   const { profile, tenant } = await requireAuthContextFromRequest(request);
   const json = await request.json();
-  const { title } = createConversationRequest.parse(json);
+  const { content } = createConversationRequest.parse(json);
+  const title = await createConversationTitle(content);
 
   const rs = await db
     .insert(schema.conversations)
@@ -81,4 +87,25 @@ export async function GET(request: NextRequest) {
     page: validatedPage,
     totalPages: Math.ceil(count / validatedLimit),
   });
+}
+
+/*
+params: content: string
+returns: string - appropriate name for this conversation
+*/
+async function createConversationTitle(content: string) {
+  const model = openai(DEFAULT_NAMING_MODEL);
+  const systemPrompt = NAMING_SYSTEM_PROMPT;
+  const userPrompt = `
+  Here is the initial message from the user:
+  ${content}
+  Please generate a title for this conversation.
+  `;
+
+  const { text } = await generateText({
+    model,
+    system: systemPrompt,
+    prompt: userPrompt,
+  });
+  return text;
 }
